+++ rosetta/tests/__init__.py
@@ -537,3 +537,22 @@ class RosettaTestCase(TestCase):
         r = self.client.post(reverse('rosetta-home'), dict(m_e48f149a8b2e8baa81b816c0edf93890='Hello, world', _next='_next'))
         r = self.client.get(reverse('rosetta-home'))
         self.assertTrue('Progress: 25.00%' in str(r.content))
+
+    def test_24_replace_access_control(self):
+        # Test default access control allows access
+        url = reverse('rosetta-home')
+        response = self.client.get(url)
+        self.assertEqual(200, response.status_code)
+
+        # Now replace access control, and check we get redirected
+        settings.ROSETTA_ACCESS_CONTROL_FUNCTION = 'rosetta.tests.no_access'
+        response = self.client.get(url)
+        self.assertEqual(302, response.status_code)
+
+        # Restore setting to default
+        settings.ROSETTA_ACCESS_CONTROL_FUNCTION = None
+
+
+# Stubbed access control function
+def no_access(user):
+    return False
